backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pyjwt==2.9.0
redis==5.0.8
opencv-python-headless==4.10.0.84
ultralytics==8.3.11
numpy==1.26.4
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-socketio==5.11.3
